requirements.txt
streamlit
fastapi
uvicorn
chromadb
sentence-transformers
pandas
requests
python-dotenv
vaderSentiment
plotly
torch
transformers
